+++ src/components/AlleredeSvart.tsx
