+++ src/pages/Sporsmal.tsx
@@ -1,12 +1,12 @@
 import React, {useEffect, useState} from 'react';
 import { BrowserRouter, Route } from 'react-router-dom';
-import OnsketMoteFormSporsmal from './page/onsket-moteform/OnsketMoteFormSporsmal';
-import HvaMotetSkalHandleOmSporsmal, { PAGE_ID as HVA_PAGE_ID } from './page/hva-skal-mote-handle-om/HvaMotetSkalHandleOmSporsmal';
-import NarPasserMotetSporsmal, { PAGE_ID as NAR_PAGE_ID } from './page/nar-passer-motet/NarPasserMotetSporsmal';
-import DinSituasjonSporsmal, { PAGE_ID as SITUASJON_PAGE_ID } from './page/din-situasjon/DinSituasjonSporsmal';
-import Oppsummering, { PAGE_ID as OPPSUMMERING_PAGE_ID } from './page/oppsummering/Oppsummering';
+import OnsketMoteFormSporsmal from './onsket-moteform/OnsketMoteFormSporsmal';
+import HvaMotetSkalHandleOmSporsmal, { PAGE_ID as HVA_PAGE_ID } from './hva-skal-mote-handle-om/HvaMotetSkalHandleOmSporsmal';
+import NarPasserMotetSporsmal, { PAGE_ID as NAR_PAGE_ID } from './nar-passer-motet/NarPasserMotetSporsmal';
+import DinSituasjonSporsmal, { PAGE_ID as SITUASJON_PAGE_ID } from './din-situasjon/DinSituasjonSporsmal';
+import Oppsummering, { PAGE_ID as OPPSUMMERING_PAGE_ID } from './oppsummering/Oppsummering';
 import './Sporsmal.less';
-import PageChangeListener from '../components/pange-change-listener/PageChangeListener';
+import PageChangeListener from '../components/PageChangeListener';
 import {getOppfolging} from "../components/api/api";
 import {OppfolgingData} from "../components/api/dataTypes";
 import {AlertStripeAdvarsel} from "nav-frontend-alertstriper";
