+++ src/pages/din-situasjon/DinSituasjonSporsmal.tsx
@@ -1,14 +1,14 @@
 import React, { useReducer } from 'react';
-import NarPasserMotetView, { SPORSMAL } from './NarPasserMotetView';
-import { fetchReducer, initialFetchState } from '../../fetchReducer';
-import { dispatchDialogData } from '../../dispatchDialogData';
-import { PAGE_ID as OPPSUMMERING_PAGE_ID } from '../oppsummering/Oppsummering';
 import { RouteComponentProps, withRouter } from 'react-router-dom';
-import { getQueryParam } from '../../../components/util/querystring-utils';
+import DinSituasjonView, { SPORSMAL } from './DinSituasjonView';
+import { fetchReducer, initialFetchState } from '../fetchReducer';
+import { dispatchDialogData } from '../dispatchDialogData';
+import { PAGE_ID as OPPSUMMERING_PAGE_ID } from '../oppsummering/Oppsummering';
+import { getQueryParam } from '../../components/util/querystring-utils';
 
-export const PAGE_ID = 'nar-passer-motet';
+export const PAGE_ID = 'din-situasjon';
 
-function NarPasserMotetSporsmal(props: RouteComponentProps) {
+function DinSituasjonSporsmal(props: RouteComponentProps) {
     const [fetchState, fetchDispatch] = useReducer(fetchReducer, initialFetchState);
     const dialogId = getQueryParam(props.location.search, 'dialogId');
     const answered = getQueryParam(props.location.search, 'answered') === 'true';
@@ -19,7 +19,7 @@ function NarPasserMotetSporsmal(props: RouteComponentProps) {
             dispatchDialogData(dialogInputData, fetchDispatch)
                 .then((res) => {
                     props.history.replace(props.location.pathname + `?dialogId=${res.id}&answered=true`);
-                    props.history.push(`/${OPPSUMMERING_PAGE_ID}?dialogId=${res.id}`);
+                    props.history.push(`/${OPPSUMMERING_PAGE_ID}`);
                 });
         } else {
             const queryParam = dialogId ? `?dialogId=${dialogId}` : '';
@@ -28,7 +28,7 @@ function NarPasserMotetSporsmal(props: RouteComponentProps) {
     };
 
     return (
-        <NarPasserMotetView
+        <DinSituasjonView
             onSubmit={onSubmit}
             loading={fetchState.loading}
             answered={answered}
@@ -36,4 +36,4 @@ function NarPasserMotetSporsmal(props: RouteComponentProps) {
     );
 }
 
-export default withRouter(NarPasserMotetSporsmal);
+export default withRouter(DinSituasjonSporsmal);
