+++ src/pages/nar-passer-motet/NarPasserMotetSporsmal.tsx
@@ -1,14 +1,14 @@
 import React, { useReducer } from 'react';
-import HvaMotetSkalHandleOmView, { SPORSMAL } from './HvaMotetSkalHandleOmView';
-import { fetchReducer, initialFetchState } from '../../fetchReducer';
-import { dispatchDialogData } from '../../dispatchDialogData';
+import NarPasserMotetView, { SPORSMAL } from './NarPasserMotetView';
+import { fetchReducer, initialFetchState } from '../fetchReducer';
+import { dispatchDialogData } from '../dispatchDialogData';
+import { PAGE_ID as OPPSUMMERING_PAGE_ID } from '../oppsummering/Oppsummering';
 import { RouteComponentProps, withRouter } from 'react-router-dom';
-import { PAGE_ID as NAR_PAGE_ID } from '../nar-passer-motet/NarPasserMotetSporsmal';
-import { getQueryParam } from '../../../components/util/querystring-utils';
+import { getQueryParam } from '../../components/util/querystring-utils';
 
-export const PAGE_ID = 'hva-motet-skal-handle-om';
+export const PAGE_ID = 'nar-passer-motet';
 
-function HvaMotetSkalHandleOmSporsmal(props: RouteComponentProps) {
+function NarPasserMotetSporsmal(props: RouteComponentProps) {
     const [fetchState, fetchDispatch] = useReducer(fetchReducer, initialFetchState);
     const dialogId = getQueryParam(props.location.search, 'dialogId');
     const answered = getQueryParam(props.location.search, 'answered') === 'true';
@@ -16,20 +16,19 @@ function HvaMotetSkalHandleOmSporsmal(props: RouteComponentProps) {
     const onSubmit = (value: string) => {
         if (value.length !== 0) {
             const dialogInputData = {svar: value, spm: SPORSMAL, dialogId: dialogId};
-
             dispatchDialogData(dialogInputData, fetchDispatch)
                 .then((res) => {
                     props.history.replace(props.location.pathname + `?dialogId=${res.id}&answered=true`);
-                    props.history.push(`/${NAR_PAGE_ID}?dialogId=${res.id}`);
+                    props.history.push(`/${OPPSUMMERING_PAGE_ID}?dialogId=${res.id}`);
                 });
         } else {
             const queryParam = dialogId ? `?dialogId=${dialogId}` : '';
-            props.history.push(`/${NAR_PAGE_ID}${queryParam}`);
+            props.history.push(`/${OPPSUMMERING_PAGE_ID}${queryParam}`);
         }
     };
 
     return (
-        <HvaMotetSkalHandleOmView
+        <NarPasserMotetView
             onSubmit={onSubmit}
             loading={fetchState.loading}
             answered={answered}
@@ -37,4 +36,4 @@ function HvaMotetSkalHandleOmSporsmal(props: RouteComponentProps) {
     );
 }
 
-export default withRouter(HvaMotetSkalHandleOmSporsmal);
+export default withRouter(NarPasserMotetSporsmal);
