+++ src/pages/onsket-moteform/OnsketMoteFormSporsmal.tsx
@@ -1,13 +1,13 @@
 import React, { useReducer } from 'react';
 import { RouteComponentProps, withRouter } from 'react-router-dom';
 import OnsketMoteFormView, { MoteForm, moteFormValue, SPORSMAL, WRITE } from './OnsketMoteFormView';
-import { fetchReducer, initialFetchState } from '../../fetchReducer';
-import { frontendLogger } from '../../../components/util/frontendlogger';
-import { dispatchDialogData } from '../../dispatchDialogData';
+import { fetchReducer, initialFetchState } from '../fetchReducer';
+import { frontendLogger } from '../../components/util/frontendlogger';
+import { dispatchDialogData } from '../dispatchDialogData';
 import { PAGE_ID as DIN_SITUASJON_PAGE_ID } from '../din-situasjon/DinSituasjonSporsmal';
 import { PAGE_ID as HANDLER_OM_PAGE_ID } from '../hva-skal-mote-handle-om/HvaMotetSkalHandleOmSporsmal';
-import { dispatchMotestotte } from '../../dispatchMotestotteData';
-import { getQueryParam } from '../../../components/util/querystring-utils';
+import { dispatchMotestotte } from '../dispatchMotestotteData';
+import { getQueryParam } from '../../components/util/querystring-utils';
 
 export const PAGE_ID = 'onsket-mote-form';
